+++ src/main/java/data/duke.txt
@@ -1,1 +1,4 @@
-T | 0 | read
+T | 0 | read book
+D | 0 | deadline return book | 2023-11-11T18:00
+E | 1 | event proj meeting | 2023-11-11T16:00 - 2023-11-12T17:00
+E | 0 |  | 2023-12-11T12:00 - 2023-12-11T15:00
